move test to cl_ch module, now that we are implementing our own version
